requirements.txt
# Core Dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.11.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=2.15.0

# Data Processing
pandas>=2.0.0
imutils>=0.5.4

# Audio Alerts
pygame>=2.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
